add: Tratamento de exceptions, model user, relacionamentos entre models, ajuste geral do crud

+++ src/main/java/br/edu/utfpr/dungeontable/service/ItemService.java
@@ -1,5 +1,7 @@
 package br.edu.utfpr.dungeontable.service;
 
+import br.edu.utfpr.dungeontable.exception.BusinessException;
+import br.edu.utfpr.dungeontable.exception.ErrorCode;
 import br.edu.utfpr.dungeontable.model.tools.Item;
 import br.edu.utfpr.dungeontable.repository.ItemRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,11 +19,27 @@ public class ItemService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     public Item save(Item item) {
+        if (item.getName() == null || item.getName().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "name item");
+        } else if (item.getCategory() == null || item.getCategory().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "category item");
+        } else if (item.getPrice() == null) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "price item");
+        }
         return itemRepository.save(item);
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
     public Item update(Item item) {
+        if (item.getId() == null) {
+            throw new BusinessException(ErrorCode.ID_REQUIRED);
+        } else if (item.getName() == null || item.getName().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "name item");
+        } else if (item.getCategory() == null || item.getCategory().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "category item");
+        } else if (item.getPrice() == null) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "price item");
+        }
         return itemRepository.save(item);
     }
 
@@ -35,6 +53,9 @@ public class ItemService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     public void delete(Long id) {
+        if(id == null){
+            throw new BusinessException(ErrorCode.ID_REQUIRED);
+        }
         itemRepository.deleteById(id);
     }
 }
